Clean up of panaosic decompression
Former-commit-id: 39ef7064238902b3ff8627e7b0cb4a9a1756b8c0
Former-commit-id: 0e1886541bea3871afa67904223474198971380b [formerly 025538f7998b22490891e63443a7dade28904e41]
Former-commit-id: 27710c736b75e58eef29cdc51f190814a5e9797a
Former-commit-id: b7b93a4894efb2b2b5d6b513cc78a502350c3fe4

+++ Source/RawParser/Model/Parser/Decoder/RawDecoder.cs
@@ -54,9 +54,9 @@ namespace RawNet
         public abstract void DecodeMetadata();
 
         /* This is faster - at least when compiled on visual studio 32 bits */
-        protected int Other_abs(int x)
+        protected long Other_abs(long x)
         {
-            int mask = x >> 31;
+            long mask = x >> 31;
             return (x + mask) ^ mask;
         }
     }
